[#180] feat : 서버 사이드화 및 edit 페이지 get 데이터 뿌리기 적용
