+++ app/_components/BackHeader/style.css.ts
@@ -5,7 +5,6 @@ export const header = style({
   padding: "1.55rem 1.6rem",
 
   position: "fixed",
-  top: "5.6rem",
 
   width: "100%",
