+++ app/_constants/pagesWithHeader.ts
@@ -1,4 +1,4 @@
 export const pagesWithHeader = {
   logo: ["/home/select", "/healthlog/select", "/diary/select", "/settings/faq", "/settings/notice", "/settings/ask", "/settings", "/settings/received-invite"],
-  petGroup: ["/home", "/healthlog", "/diary", "/healthlog/edit", "/healthlog/create", "/diary/edit", "/settings/member", "/settings/invitation"],
+  petGroup: ["/home", "/healthlog", "/diary", "/settings/member", "/settings/invitation"],
 };
