+++ app/healthlog/edit/[id]/page.tsx
@@ -1,208 +1,11 @@
-"use client";
+import EditForm from "@/app/healthlog/_components/EditForm";
+import { cookies } from "next/headers";
 
-import { getLogDetail, putLogs } from "@/app/_api/log";
-import BackHeader from "@/app/_components/BackHeader";
-import DateInput from "@/app/_components/DateInput";
-import { LogDetailType } from "@/app/_types/log/types";
-import convertTime12to24 from "@/app/_utils/convertTime12to24";
-import SelectMateDropdown from "@/app/healthlog/_components/SelectMateDropdown";
-import SubtypeDetail from "@/app/healthlog/_components/SubtypeDetail";
-import { buttonTypes } from "@/public/data/buttonTypes";
-import { subtypeOptions } from "@/public/data/subtypeOptions";
-import { useQuery } from "@tanstack/react-query";
-import { usePathname } from "next/navigation";
-import { useEffect, useRef, useState } from "react";
-import { FieldValues, useForm } from "react-hook-form";
-import * as styles from "./page.css";
+const EditPage = ({ params: { id } }: { params: { id: string } }) => {
+  const logId = Number(id);
+  const petId = Number(cookies().get("petId")?.value);
 
-const Page = () => {
-  const [visibleSubtype, setVisibleSubtype] = useState<keyof typeof subtypeOptions | "CUSTOM" | "WALK" | null>("FEED");
-  const [selectedType, setSelectedType] = useState<string>("FEED");
-  const [kakaoLocationId, setKakaoLocationId] = useState<number | null>(null);
-  const [activeButtonGroup, setActiveButtonGroup] = useState("");
-  const [selectedGuardianId, setSelectedGuardianId] = useState<string>("");
-  const topSubtypeRef = useRef<HTMLDivElement>(null);
-  const bottomSubtypeRef = useRef<HTMLDivElement>(null);
-
-  const pathname = usePathname();
-  const segments = pathname.split("/");
-  const logId = Number(segments[segments.length - 1]);
-  const petId = Number(localStorage.getItem("petId"));
-
-  const { data: logDetailData, error } = useQuery<LogDetailType, Error>({
-    queryKey: ["LogDetail", petId, logId],
-    queryFn: () => getLogDetail(Number(petId), Number(logId)),
-    enabled: !!petId,
-  });
-
-  const {
-    handleSubmit,
-    register,
-    setValue,
-    getValues,
-    watch,
-    formState: { errors },
-  } = useForm<FieldValues>();
-
-  const handleTypeButtonClick = (type: string, group: string) => {
-    if (selectedType === type && activeButtonGroup === group) {
-      setVisibleSubtype(type as keyof typeof subtypeOptions | "CUSTOM" | "WALK");
-      setActiveButtonGroup("");
-    } else {
-      setSelectedType(type);
-      setVisibleSubtype(type as keyof typeof subtypeOptions | "CUSTOM" | "WALK");
-      setActiveButtonGroup(group);
-    }
-  };
-
-  const handleLocationSelect = (id: number | null) => {
-    setKakaoLocationId(id);
-  };
-
-  const handleKeyPress = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter" && !e.shiftKey) {
-      e.preventDefault();
-    }
-  };
-
-  const onSubmit = (data: any) => {
-    const date = data.date;
-    const time = convertTime12to24(data.time);
-    const datetime = `${date}T${time}`;
-
-    const logData = {
-      type: selectedType === "CUSTOM" ? data.type : selectedType,
-      subType: data.subtype,
-      isCustomLocation: selectedType === "WALK",
-      kakaoLocationId: selectedType === "WALK" ? kakaoLocationId : null,
-      datetime: datetime,
-      isComplete: true,
-      isImportant: data.isImportant,
-      memo: data.memo,
-      managerId: selectedGuardianId,
-    };
-
-    console.log(petId, logId, logData);
-    putLogs(petId, logId, logData);
-  };
-
-  useEffect(() => {
-    if (logDetailData) {
-      setValue("date", logDetailData.date);
-      setValue("time", logDetailData.time);
-      setValue("type", logDetailData.type);
-      setValue("subType", logDetailData.subType);
-      setValue("memo", logDetailData.memo);
-      setValue("isImportant", logDetailData.isImportant);
-    }
-  }, [logDetailData, setValue]);
-
-  useEffect(() => {
-    setValue("memo", "");
-    setValue("subtype", "");
-    setValue("isImportant", false);
-  }, [visibleSubtype, setValue]);
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (event.target instanceof HTMLElement) {
-        if (event.target.id === "submit-button") {
-          return;
-        }
-
-        if (topSubtypeRef.current && !topSubtypeRef.current.contains(event.target) && activeButtonGroup === "top") {
-          setVisibleSubtype(null);
-          setActiveButtonGroup("");
-        }
-        if (bottomSubtypeRef.current && !bottomSubtypeRef.current.contains(event.target) && activeButtonGroup === "bottom") {
-          setVisibleSubtype(null);
-          setActiveButtonGroup("");
-        }
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [activeButtonGroup]);
-
-  if (error) return <div>Error occurred!</div>;
-
-  return (
-    <>
-      <BackHeader title="건강수첩 수정하기" />
-      <div className={styles.container}>
-        <form onSubmit={handleSubmit(onSubmit)} className={styles.formItems} onKeyDown={handleKeyPress}>
-          <div className={styles.inputWrapper}>
-            <DateInput register={register} setValue={setValue} getValue={getValues} />
-          </div>
-          <div className={styles.inputWrapper}>
-            <label>담당 메이트</label>
-            <SelectMateDropdown
-              onSelect={(guardianId) => {
-                setSelectedGuardianId(guardianId);
-              }}
-            />
-          </div>
-
-          <div className={styles.inputWrapper}>
-            <div className={styles.buttonGroup}>
-              {buttonTypes.slice(0, 3).map(({ type, label }) => (
-                <button
-                  key={type}
-                  className={`${styles.typeButton} ${selectedType === type ? styles.typeButtonSelected : ""}`}
-                  type="button"
-                  onClick={() => {
-                    handleTypeButtonClick(type, "top");
-                    setSelectedType(type);
-                  }}
-                >
-                  {label}
-                  <span className={selectedType === type ? styles.addIconSelected : styles.addIcon}>+</span>
-                </button>
-              ))}
-            </div>
-            <div ref={topSubtypeRef}>
-              {visibleSubtype && activeButtonGroup === "top" && (
-                <SubtypeDetail visibleSubtype={visibleSubtype} register={register} watch={watch} errors={errors} setValue={setValue} onLocationSelect={handleLocationSelect} />
-              )}
-            </div>
-          </div>
-
-          <div className={styles.inputWrapper}>
-            <div className={styles.buttonGroup}>
-              {buttonTypes.slice(3).map(({ type, label }) => (
-                <button
-                  key={type}
-                  className={`${styles.typeButton} ${selectedType === type ? styles.typeButtonSelected : ""}`}
-                  type="button"
-                  onClick={() => {
-                    handleTypeButtonClick(type, "bottom");
-                    setSelectedType(type);
-                  }}
-                >
-                  {label}
-                  <span className={selectedType === type ? styles.addIconSelected : styles.addIcon}>+</span>
-                </button>
-              ))}
-            </div>
-            <div ref={bottomSubtypeRef}>
-              {visibleSubtype && activeButtonGroup === "bottom" && (
-                <SubtypeDetail visibleSubtype={visibleSubtype} register={register} watch={watch} errors={errors} setValue={setValue} onLocationSelect={handleLocationSelect} />
-              )}
-            </div>
-          </div>
-
-          <div>
-            <button className={styles.submitButton} type="submit" id="submit-button">
-              저장
-            </button>
-          </div>
-        </form>
-      </div>
-    </>
-  );
+  return <EditForm petId={petId} logId={logId} />;
 };
 
-export default Page;
+export default EditPage;
